# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GadgetByteNepal",
        "description": "GadgetByte Nepal Rest Apis Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/post-image": {
            "post": {
                "tags": [
                    "Post Image"
                ],
                "summary": "Create a new Post Image",
                "operationId": "f95920a9ce6fe8109312d541ae4ba03e",
                "parameters": [
                    {
                        "name": "X-User-Id",
                        "in": "header",
                        "description": "User ID for authentication",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Post Image data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StorePostImageRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully created image",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostImageResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity (Validation error)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    },
                    {
                        "X-User-Id": []
                    }
                ]
            }
        },
        "/api/v1/post-image/{id}": {
            "delete": {
                "tags": [
                    "Post Image"
                ],
                "summary": "Delete an post image",
                "description": "Remove the specified resource from storage.",
                "operationId": "f6073d4b3fee9f1432ff960e9b8a12c0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the post image to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "X-User-Id",
                        "in": "header",
                        "description": "User ID for authentication",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted post image",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Post Image deleted successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PostImageResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    },
                    {
                        "X-User-Id": []
                    }
                ]
            }
        },
        "/api/v1/post-price": {
            "post": {
                "tags": [
                    "Post Price"
                ],
                "summary": "Create a new Post Price",
                "operationId": "ffc5b8610c6845b8389221961f023db2",
                "parameters": [
                    {
                        "name": "X-User-Id",
                        "in": "header",
                        "description": "User ID for authentication",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Post Price data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StorePostPriceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully created post",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostPriceResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity (Validation error)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    },
                    {
                        "X-User-Id": []
                    }
                ]
            }
        },
        "/api/v1/post-price/{id}": {
            "put": {
                "tags": [
                    "Post Price"
                ],
                "summary": "Update an existing post price",
                "description": "Update the specified resource in storage.",
                "operationId": "3615ff98311dc3b29cf522b01aa7a53a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the post price to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "X-User-Id",
                        "in": "header",
                        "description": "User ID for authentication",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Post Price data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePostPriceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully updated post price",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostPriceResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity (Validation error)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    },
                    {
                        "X-User-Id": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Post Price"
                ],
                "summary": "Delete an post price",
                "description": "Remove the specified resource from storage.",
                "operationId": "4f954f87073fc974504075b379c1991d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the post price to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "X-User-Id",
                        "in": "header",
                        "description": "User ID for authentication",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted post price",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Post Price deleted successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PostPriceResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    },
                    {
                        "X-User-Id": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ValidationErrorResponse": {
                "title": "Validation Error Response",
                "description": "An object representing a validation error response",
                "properties": {
                    "message": {
                        "description": "The error message",
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "description": "An object containing field validation error messages. (This may varies from your api endpoints.)",
                        "type": "object",
                        "example": {
                            "postkey": [
                                "The postkey field is required."
                            ]
                        }
                    }
                },
                "type": "object"
            },
...
            "StorePostImageRequest": {
                "title": "Store Post Image Request",
                "description": "Request data for storing post images",
                "required": [
                    "post_id",
                    "slug",
                    "images"
                ],
                "properties": {
                    "post_id": {
                        "description": "The ID of the post",
                        "type": "integer"
                    },
                    "slug": {
                        "description": "The unique images slug for post image",
                        "type": "string"
                    },
                    "images": {
                        "description": "Array of post image details",
                        "type": "array",
                        "items": {
                            "title": "Post Image",
                            "description": "Post image details",
                            "required": [
                                "title",
                                "amount"
                            ],
                            "properties": {
                                "title": {
                                    "description": "The title of the post",
                                    "type": "string",
                                    "example": "Samsung Galaxy S21 Pro"
                                },
                                "image": {
                                    "description": "The image path",
                                    "type": "number",
                                    "format": "string",
                                    "example": "/post/post.png"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "StorePostPriceRequest": {
                "title": "Store Post Price Request",
                "description": "Request data for storing post prices",
                "required": [
                    "post_id",
                    "price"
                ],
                "properties": {
                    "post_id": {
                        "description": "The ID of the post",
                        "type": "integer"
                    },
                    "price": {
                        "description": "Array of post price details",
                        "type": "array",
                        "items": {
                            "title": "Post Price",
                            "description": "Post price details",
                            "required": [
                                "title",
                                "amount"
                            ],
                            "properties": {
                                "title": {
                                    "description": "The title of the post",
                                    "type": "string",
                                    "example": "Samsung Galaxy S21 Pro"
                                },
                                "amount": {
                                    "description": "The price amount",
                                    "type": "number",
                                    "format": "float",
                                    "example": "10000.00"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "UpdatePostPriceRequest": {
                "title": "Post Price  Update Request",
                "description": "Post Price  update request data",
                "required": [
                    "title",
                    "price"
                ],
                "properties": {
                    "title": {
                        "description": "The Post Price display name",
                        "type": "string",
                        "example": "Samsung galaxy S23 Ultra"
                    },
                    "price": {
                        "description": "The Amount of the  PostPrice.",
                        "type": "number",
                        "format": "float",
                        "example": "10000"
                    }
                },
                "type": "object"
            },
...
            "PostImageResource": {
                "title": "Post Image Resource",
                "description": "Post Image resource representation",
                "properties": {
                    "id": {
                        "description": "The unique id of the post image",
                        "type": "integer",
                        "example": "1"
                    },
                    "title": {
                        "description": "The title of the post image",
                        "type": "string",
                        "example": "Samsung Galaxy S23"
                    },
                    "slug": {
                        "description": "The slug of the post image",
                        "type": "string",
                        "example": "Samsung Galaxy S23"
                    },
                    "post_id": {
                        "description": "The Id of the post ",
                        "type": "integer",
                        "example": "1"
                    },
                    "image": {
                        "description": "The relative path of the post image",
                        "type": "string",
                        "example": "/post/post.png"
                    }
                },
                "type": "object"
            },
            "PostPriceResource": {
                "title": "Post Price Resource",
                "description": "Post Price resource representation",
                "properties": {
                    "id": {
                        "description": "The unique id of the post price",
                        "type": "integer",
                        "example": "1"
                    },
                    "title": {
                        "description": "The title of the post price",
                        "type": "string",
                        "example": "Samsung Galaxy S23"
                    },
                    "post_id": {
                        "description": "The Id of the post ",
                        "type": "integer",
                        "example": "1"
                    },
                    "price": {
                        "description": "The amount of the post price",
                        "type": "number",
                        "format": "float",
                        "example": "10000"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "description": "Bearer Token Authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}